move AsyncTask to Engine package and hide waitForReadyToPlay
